cv-cctv-action-timestamping/requirements.txt
streamlit==1.51.0  
opencv-python==4.8.1.78  
numpy==1.25.3  
tensorflow==2.16.0  
keras==2.16.0  
Pillow==10.0.1  
matplotlib==3.8.0  
tqdm==4.66.1  
joblib==1.3.2  
sentence-transformers==5.1.2  
torch==2.1.0